Removes `CreateAnalysis` API call
